Many changes.
The structure of the original radio does not fit very well with the idea
of navigating through the “directories” of mpd. A further incompatibility
between the old code and the new code is that the old radio class has
been designed to be a container for structural programming while the
new LCD code is more object oriented. This made the program inconsistent
and unstable.